requirements.txt
streamlit
numpy
pillow
python-dotenv
google-generativeai
